+++ src/ImageViewer/image_viewer.tsx
@@ -39,14 +39,31 @@ export default function ImageViewer(props: {
     >
     </div>
 
-    if (mode === 'full screen') return <Modal onClose={() => { if (props.onClose) props.onClose() }}>
-        <div className={`vieolo-image-viewer vieolo-image-viewer--full-screen`}>
-            {frame}
-            <div className="vieolo-image-viewer__container">
-                {content}
+    if (mode === 'full screen') {
+        return <Modal onClose={() => { 
+			// Here, we are handling the browser back button.
+			// Since the Viewer is full screen, the only way to close the `Modal` is to
+			// click the back button
+			// 
+			// If the viewer started as `embedded` and has been set to `full screen` by the user,
+			// we set the mode to `embedded`
+			//
+			// Else if the viewer start as `full screen` (which means will remain `full screen` in its lifetime)
+			// and there is a a `onClose` callback, we trigger the `onClose`
+			if (props.context === 'embedded' && mode === 'full screen') {
+				setMode('embedded');
+			} else if (props.context === 'full screen' && mode === 'full screen' && props.onClose) {
+				props.onClose()
+			}
+        }}>
+            <div className={`vieolo-image-viewer vieolo-image-viewer--full-screen`}>
+                {frame}
+                <div className="vieolo-image-viewer__container">
+                    {content}
+                </div>
             </div>
-        </div>
-    </Modal>
+        </Modal>
+    }
 
     return <div className={`vieolo-image-viewer vieolo-image-viewer--embedded`}>
         {frame}